Working for GPT blocking permissions
