+++ massgen/mcp_tools/__init__.py
@@ -43,6 +43,11 @@ from .backend_utils import (
 )
 from .converters import MCPConverters, MCPFormatValidator
 
+# Hook system for function call interception
+from .hooks import HookType, FunctionHook, PermissionHook, HookResult
+from .hook_manager import FunctionHookManager, function_hook_manager
+from .permission_bridge import PathPermissionBridge, FilesystemManagerBridge, setup_backend_permission_hooks
+
 __all__ = [
     # Core client classes
     "MCPClient",
@@ -83,4 +88,15 @@ __all__ = [
     "MCPExecutionManager",
     "MCPConverters",
     "MCPFormatValidator",
+    # Hook system
+    "HookType",
+    "FunctionHook",
+    "PermissionHook",
+    "HookResult",
+    "FunctionHookManager",
+    "function_hook_manager",
+    # Permission bridge system
+    "PathPermissionBridge",
+    "FilesystemManagerBridge",
+    "setup_backend_permission_hooks",
 ]
