+++ massgen/mcp_tools/permission_wrapper.py
